requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
pandas>=2.2.3
openpyxl==3.1.5
scikit-learn==1.5.2
numpy>=2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pdfplumber==0.11.0
bcrypt==4.1.3
PyJWT==2.8.0
APScheduler==3.10.4